+++ xyui/avalonia/gallery/XYUI.Avalonia.Gallery/Views/PaletteView.axaml.cs
@@ -1,0 +1,12 @@
+using Avalonia.Controls;
+
+namespace XYUI.Avalonia.Gallery.Views;
+
+public partial class PaletteView : UserControl
+{
+    public PaletteView()
+    {
+        InitializeComponent();
+        DataContext = PaletteViewModel.Create();
+    }
+}